Imported Upstream version 1.532.1

+++ core/src/main/java/hudson/model/Run.java
@@ -52,14 +52,12 @@ import hudson.security.AccessControlled;
 import hudson.security.Permission;
 import hudson.security.PermissionGroup;
 import hudson.tasks.BuildWrapper;
-import hudson.tasks.BuildStep;
 import hudson.tasks.test.AbstractTestResultAction;
 import hudson.util.FlushProofOutputStream;
 import hudson.util.FormApply;
 import hudson.util.IOException2;
 import hudson.util.LogTaskListener;
 import hudson.util.XStream2;
-import hudson.util.ProcessTree;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -117,11 +115,17 @@ import org.kohsuke.stapler.interceptor.RequirePOST;
 import java.io.FileOutputStream;
 import java.io.OutputStream;
 
+import java.io.StringWriter;
 import static java.util.logging.Level.*;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
+import jenkins.model.ArtifactManager;
+import jenkins.model.ArtifactManagerConfiguration;
+import jenkins.model.ArtifactManagerFactory;
 import jenkins.model.PeepholePermalink;
+import jenkins.model.StandardArtifactManager;
 import jenkins.model.RunAction2;
+import jenkins.util.VirtualFile;
 
 /**
  * A particular execution of {@link Job}.
@@ -189,7 +193,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
 
     /**
      * The build result.
-     * This value may change while the state is in {@link State#BUILDING}.
+     * This value may change while the state is in {@link Run.State#BUILDING}.
      */
     protected volatile Result result;
 
@@ -256,8 +260,14 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
      */
     private volatile transient RunExecution runner;
 
+    /**
+     * Artifact manager associated with this build, if any.
+     * @since 1.532
+     */
+    private @CheckForNull ArtifactManager artifactManager;
+
     private static final SimpleDateFormat CANONICAL_ID_FORMATTER = new SimpleDateFormat("yyyy-MM-dd_HH-mm-ss");
-    protected static final ThreadLocal<SimpleDateFormat> ID_FORMATTER = new IDFormatterProvider();
+    public static final ThreadLocal<SimpleDateFormat> ID_FORMATTER = new IDFormatterProvider();
     private static final class IDFormatterProvider extends ThreadLocal<SimpleDateFormat> {
                 @Override
                 protected SimpleDateFormat initialValue() {
@@ -325,6 +335,9 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
                 ((RunAction) a).onLoad();
             }
         }
+        if (artifactManager != null) {
+            artifactManager.onLoad(this);
+        }
     }
     
     /**
@@ -336,6 +349,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         List<Action> actions = new ArrayList<Action>();
         for (TransientBuildActionFactory factory: TransientBuildActionFactory.all()) {
             actions.addAll(factory.createFor(this));
+            assert !actions.contains(null) : "null action added by " + factory;
         }
         return Collections.unmodifiableList(actions);
     }
@@ -398,9 +412,10 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
      * <p>
      * When a build is {@link #isBuilding() in progress}, this method
      * returns an intermediate result.
+     * @return The status of the build, if it has completed or some build step has set a status; may be null if the build is ongoing.
      */
     @Exported
-    public Result getResult() {
+    public @CheckForNull Result getResult() {
         return result;
     }
 
@@ -463,8 +478,8 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
      * 
      * This method looks for {@link Executor} who's {@linkplain Executor#getCurrentExecutable() assigned to this build},
      * and because of that this might not be necessarily in sync with the return value of {@link #isBuilding()} &mdash;
-     * an executor holds on to {@lnk Run} some more time even after the build is finished (for example to
-     * perform {@linkplain State#POST_PRODUCTION post-production processing}.)
+     * an executor holds on to {@link Run} some more time even after the build is finished (for example to
+     * perform {@linkplain Run.State#POST_PRODUCTION post-production processing}.)
      */
     @Exported 
     public @CheckForNull Executor getExecutor() {
@@ -705,7 +720,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         BallColor baseColor;
         RunT pb = getPreviousBuild();
         if(pb==null)
-            baseColor = BallColor.GREY;
+            baseColor = BallColor.NOTBUILT;
         else
             baseColor = pb.getIconColor();
 
@@ -962,8 +977,52 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
     }
 
     /**
-     * Gets the directory where the artifacts are archived.
+     * Gets an object responsible for storing and retrieving build artifacts.
+     * If {@link #pickArtifactManager} has previously been called on this build,
+     * and a nondefault manager selected, that will be returned.
+     * Otherwise (including if we are loading a historical build created prior to this feature) {@link StandardArtifactManager} is used.
+     * <p>This method should be used when existing artifacts are to be loaded, displayed, or removed.
+     * If adding artifacts, use {@link #pickArtifactManager} instead.
+     * @return an appropriate artifact manager
+     * @since 1.532
      */
+    public final @Nonnull ArtifactManager getArtifactManager() {
+        return artifactManager != null ? artifactManager : new StandardArtifactManager(this);
+    }
+
+    /**
+     * Selects an object responsible for storing and retrieving build artifacts.
+     * The first time this is called on a running build, {@link ArtifactManagerConfiguration} is checked
+     * to see if one will handle this build.
+     * If so, that manager is saved in the build and it will be used henceforth.
+     * If no manager claimed the build, {@link StandardArtifactManager} is used.
+     * <p>This method should be used when a build step expects to archive some artifacts.
+     * If only displaying existing artifacts, use {@link #getArtifactManager} instead.
+     * @return an appropriate artifact manager
+     * @throws IOException if a custom manager was selected but the selection could not be saved
+     * @since 1.532
+     */
+    public final synchronized @Nonnull ArtifactManager pickArtifactManager() throws IOException {
+        if (artifactManager != null) {
+            return artifactManager;
+        } else {
+            for (ArtifactManagerFactory f : ArtifactManagerConfiguration.get().getArtifactManagerFactories()) {
+                ArtifactManager mgr = f.managerFor(this);
+                if (mgr != null) {
+                    artifactManager = mgr;
+                    save();
+                    return mgr;
+                }
+            }
+            return new StandardArtifactManager(this);
+        }
+    }
+
+    /**
+     * Gets the directory where the artifacts are archived.
+     * @deprecated Should only be used from {@link StandardArtifactManager} or subclasses.
+     */
+    @Deprecated
     public File getArtifactsDir() {
         return new File(getRootDir(),"archive");
     }
@@ -981,7 +1040,11 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
      */
     public List<Artifact> getArtifactsUpTo(int n) {
         ArtifactList r = new ArtifactList();
-        addArtifacts(getArtifactsDir(),"","",r,null,n);
+        try {
+            addArtifacts(getArtifactManager().root(), "", "", r, null, n);
+        } catch (IOException x) {
+            LOGGER.log(Level.WARNING, null, x);
+        }
         r.computeDisplayName();
         return r;
     }
@@ -996,18 +1059,17 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         return !getArtifactsUpTo(1).isEmpty();
     }
 
-    private int addArtifacts( File dir, String path, String pathHref, ArtifactList r, Artifact parent, int upTo ) {
-        String[] children = dir.list();
-        if(children==null)  return 0;
-        Arrays.sort(children, String.CASE_INSENSITIVE_ORDER);
+    private int addArtifacts(VirtualFile dir, String path, String pathHref, ArtifactList r, Artifact parent, int upTo) throws IOException {
+        VirtualFile[] kids = dir.list();
+        Arrays.sort(kids);
 
         int n = 0;
-        for (String child : children) {
+        for (VirtualFile sub : kids) {
+            String child = sub.getName();
             String childPath = path + child;
             String childHref = pathHref + Util.rawEncode(child);
-            File sub = new File(dir, child);
             String length = sub.isFile() ? String.valueOf(sub.length()) : "";
-            boolean collapsed = (children.length==1 && parent!=null);
+            boolean collapsed = (kids.length==1 && parent!=null);
             Artifact a;
             if (collapsed) {
                 // Collapse single items into parent node where possible:
@@ -1135,7 +1197,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
     @ExportedBean
     public class Artifact {
         /**
-         * Relative path name from {@link Run#getArtifactsDir()}
+         * Relative path name from artifacts root.
          */
     	@Exported(visibility=3)
         public final String relativePath;
@@ -1178,7 +1240,9 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
 
         /**
          * Gets the artifact file.
+         * @deprecated May not be meaningful with custom artifact managers. Use {@link ArtifactManager#load} with {@link #relativePath} instead.
          */
+        @Deprecated
         public File getFile() {
             return new File(getArtifactsDir(),relativePath);
         }
@@ -1364,12 +1428,15 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         link.delete();
 
         File rootDir = getRootDir();
+        if (!rootDir.isDirectory()) {
+            throw new IOException(rootDir + " looks to have already been deleted");
+        }
         File tmp = new File(rootDir.getParentFile(),'.'+rootDir.getName());
         
         if (tmp.exists()) {
             Util.deleteRecursive(tmp);
         }
-        // XXX on Java 7 prefer: Files.move(rootDir.toPath(), tmp.toPath(), StandardCopyOption.ATOMIC_MOVE)
+        // TODO on Java 7 prefer: Files.move(rootDir.toPath(), tmp.toPath(), StandardCopyOption.ATOMIC_MOVE)
         boolean renamingSucceeded = rootDir.renameTo(tmp);
         Util.deleteRecursive(tmp);
         // some user reported that they see some left-over .xyz files in the workspace,
@@ -1399,7 +1466,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
     /**
      * @see CheckPoint#block()
      */
-    /*package*/ static void waitForCheckpoint(CheckPoint id) throws InterruptedException {
+    /*package*/ static void waitForCheckpoint(CheckPoint id, @CheckForNull BuildListener listener, @CheckForNull String waiter) throws InterruptedException {
         while(true) {
             Run b = RunnerStack.INSTANCE.peek().getBuild().getPreviousBuildInProgress();
             if(b==null)     return; // no pending earlier build
@@ -1409,7 +1476,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
                 Thread.sleep(0);
                 continue;
             }
-            if(runner.checkpoints.waitForCheckPoint(id))
+            if(runner.checkpoints.waitForCheckPoint(id, listener, waiter))
                 return; // confirmed that the previous build reached the check point
 
             // the previous build finished without ever reaching the check point. try again.
@@ -1445,13 +1512,19 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
                 notifyAll();
             }
 
-            protected synchronized boolean waitForCheckPoint(CheckPoint identifier) throws InterruptedException {
+            protected synchronized boolean waitForCheckPoint(CheckPoint identifier, @CheckForNull BuildListener listener, @CheckForNull String waiter) throws InterruptedException {
                 final Thread t = Thread.currentThread();
                 final String oldName = t.getName();
-                t.setName(oldName+" : waiting for "+identifier+" on "+getFullDisplayName());
+                t.setName(oldName + " : waiting for " + identifier + " on " + getFullDisplayName() + " from " + waiter);
                 try {
-                    while(!allDone && !checkpoints.contains(identifier))
+                    boolean first = true;
+                    while (!allDone && !checkpoints.contains(identifier)) {
+                        if (first && listener != null && waiter != null) {
+                            listener.getLogger().println(Messages.Run__is_waiting_for_a_checkpoint_on_(waiter, getFullDisplayName()));
+                        }
                         wait();
+                        first = false;
+                    }
                     return checkpoints.contains(identifier);
                 } finally {
                     t.setName(oldName);
@@ -1523,7 +1596,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
     }
 
     /**
-     * Used in {@link RunExecution#run} to indicates that a fatal error in a build
+     * Used in {@link Run.RunExecution#run} to indicates that a fatal error in a build
      * is reported to {@link BuildListener} and the build should be simply aborted
      * without further recording a stack trace.
      */
@@ -1921,7 +1994,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         if(Functions.isArtifactsPermissionEnabled()) {
           checkPermission(ARTIFACTS);
         }
-        return new DirectoryBrowserSupport(this,new FilePath(getArtifactsDir()), project.getDisplayName()+' '+getDisplayName(), "package.png", true);
+        return new DirectoryBrowserSupport(this, getArtifactManager().root(), project.getDisplayName() + ' ' + getDisplayName(), "package.png", true);
     }
 
     /**
@@ -1988,7 +2061,7 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
             // Definitely prevented.
             return false;
         }
-        // XXX may be that keepLog is on (perhaps toggler earlier) yet isKeepLog() would be true anyway.
+        // TODO may be that keepLog is on (perhaps toggler earlier) yet isKeepLog() would be true anyway.
         // In such a case this will incorrectly return true and logKeep.jelly will allow the toggle.
         // However at least then (after redirecting to the same page) the toggle button will correctly disappear.
         return true;
@@ -2029,7 +2102,16 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
             return;
         }
 
-        delete();
+        try{
+            delete();
+        }
+        catch(IOException ex){
+            StringWriter writer = new StringWriter();
+            ex.printStackTrace(new PrintWriter(writer));
+            req.setAttribute("stackTraces", writer);
+            req.getView(this, "delete-retry.jelly").forward(req, rsp);  
+            return;
+        } 
         rsp.sendRedirect2(req.getContextPath()+'/' + getParent().getUrl());
     }
 
@@ -2261,8 +2343,13 @@ public abstract class Run <JobT extends Job<JobT,RunT>,RunT extends Run<JobT,Run
         if (result == null){
             //check transient actions too
             for(Action action: getTransientActions()){
-                if(action.getUrlName().equals(token))
+                String urlName = action.getUrlName();
+                if (urlName == null) {
+                    continue;
+                }
+                if (urlName.equals(token)) {
                     return action;
+                }
             }
             // Next/Previous Build links on an action page (like /job/Abc/123/testReport)
             // will also point to same action (/job/Abc/124/testReport), but other builds
